Rename arithmetic interpreter object
Rename to keep the simple, direct implementation, but reserve the
Interpreter object name for the modular implementation.
